requirements.txt
asgiref==3.3.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crispy-bootstrap5==0.3.1
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-autoslug==1.9.8
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-taggit==1.4.0
django-tinymce==3.3.0
idna==2.10
oauthlib==3.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
